`requirements.txt`: Disallow `dipy==1.9.0` and above
This commit is to test whether there is a meaningful difference between dipy==1.8.0 and dipy==1.9.0.

It's possible that this has an effect on the stalling test, but it's also possible that the stalling
tests were entirely flukes due to model-downloading timeouts.

If this commit and the previous commit both pass, then `dipy` makes no difference. But if this
commit passes and the previous one fails, then `dipy` does make a difference.