Replace `matrix.os` with `runner.os` to more robustly detect Win OSs in GHA workflows (#4403)
## Description

This commit is a tiny follow-up fix for PR #4402. In that PR, we
relegated the `create-release.yml` workflow to only run on cron jobs.
And, because we no longer run the workflow on PRs, we were able to
expand the OSs suite without worrying about incurring additional costs.
So, `windows-2019` was added back in addition to `windows-2022`.

However, in order to determine which installer to use (win vs. unix), we
currently check the exact runner name ("windows-2022"). So, windows-2019
wasn't detected as a Windows runner, and the workflow failed for today's
cron job.

To fix this, we can use an alternative syntax that is more general
("Windows") suggested by the official docs:
https://docs.github.com/en/actions/learn-github-actions/variables#detecting-the-operating-system