Add improvements to CSA calculation for GM/WM masks (`-angle-corr-centerline`, `float32` precision fix, doc warnings) (#4409)
## Description

This PR is a follow up to my work helping @BFCombes on
https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/4367.
During my investigation, I discovered a few small issues and areas of
improvement. And, I wanted to make sure those observations were
addressed before moving on, lest the context of the issue be lost to
time.

This PR contains 3 main changes:

- A small precision fix for float32 zooms, which removes spurious digits
for uncorrected CSA values.
- We use
[`np.spacing`](https://numpy.org/doc/stable/reference/generated/numpy.spacing.html)
to figure out what decimal place we can reliably round to -- we can't
just rely on `float32` resolution because there are some complexities
there. I found [this SO
thread](https://stackoverflow.com/questions/32465481/what-exactly-is-the-resolution-parameter-of-numpy-float)
to be a helpful resource on the topic.
- New warnings in the tutorials + help description, communicating a lack
of precision for CSA values.
- A new option `-angle-corr-centerline`, which enables angle correction
to be performed for irregularly-shaped segmentations such as GM/WM
masks. This option is then patched into the tutorials, to avoid needing
to resort to `-angle-corr 0`.

## Linked issues

Fixes #4367.
Fixes #4378.

---------

Co-authored-by: Mathieu Guay-Paquet <mathieu.guay-paquet@polymtl.ca>
Co-authored-by: Sandrine Bédard <71230552+sandrinebedard@users.noreply.github.com>