Update `check_image_kind` heuristic to account for PSIR images
The primary goal of this commit is to avoid treating PSIR anat images as label images when
resampling for QC reports. There is less of a risk for treating label images as anat images, since
`sct_qc -p sct_label_vertebrae` doesn't perform resampling. (Though, `check_image_kind` is also
used for picking the colormap to use for `fsleyes` commands, so we do still need to consider
labeled segmentations.)

Given that label images are typically 0 outside the labels themselves, I focused on the
percentages of voxels equal to 0 as a heuristic to tell the two apart:

- PSIR images:  Whole values, but with a "percentage of zero voxels" ranging from 2-10%
- Label images: Whole values, but with a "percentage of zero voxels" >98% (assuming typical crop)

For example, the TotalSegmentator labels we tested in a previous PR still exceeded 98.9%
"zero voxels", despite including disc labels, vertebrae labels, cord labels and CSF labels.)

That said, my only concern are labels that are heavily cropped to exclude extraneous
tissue/space/etc. In that case, the percentage of zero voxels may dip far below 99%. So, that's
why I used a conservative "50%" threshold here. and we may want to go even lower, since the
PSIR images I tested never exceeded 10% 0 voxels.