`sct_check_dependencies`: Move optional header to separate `if:`
Currently, "optional" checks are UNIX-only. (The header won't be printed for Windows machines,
since Windows doesn't support FSLeyes in the first place.)

However, we may want to check GPU dependencies on Windows, too. So, we want to leave room for the
header to be printed for other conditions (e.g. "+cpu" not in __torch_version__).