Keep CSS/JS/etc. assets up to date when QC report is regenerated with a newer version of SCT (#4427)
When we generate QC reports, we universally update `index.html` with the
copy stored in the repo, due to the behavior described in #4423.

This can cause problems when, for example, `index.html` and `main.js`
are both updated simultaneously between SCT versions (e.g. #4317). In
such a case, if QC reports are regenerated using the newer SCT version,
`index.html` will be updated with the changes, but `main.js` will not.
And, since `index.html` and `main.js` are so interdependent, this can
often cause broken behavior in the QC report.

So, this PR ensures that the assets are updated alongside `index.html`,
but only if the contents differ.