Add comment for instances where data cannot be int64
nibabel will throw an error for instances of `Nifti1Image(data, affine)` when data is of type
int64 (and no header/dtype arguments are specified).

However, for these 3 instances of `Nifti1Image(data, affine)`, it is impossible for the array
to be of type int64 (due to both np.zeros and img.get_fdata returning float arrays). So, instead
of rewriting the calls to specify `dtype` or `header`, I've just added a little note.