Move `index.html` inside the `_assets` folder
Since my goal is to isolate the "JSON-to-HTML refresh" code as a standalone script, the usage of
'importlib' to find the template file for `index.html` is a little problematic, since it relies
on knowledge of where the `spinalcordtoolbox` package exists.

However, if we put the template file for `index.html` inside the `_assets` folder, it will be
copied along with everything else. Meaning, the standalone "refresh" script can use its own
asset folder to generate new `index.html` files/