Replace `locked_file` with `mutex` to address NFS mount issue
With the previous solution, if the destination `index.html` file was on an NFS-mounted
drive, we would trigger an issue inherent to portalocker where locks would fail.

But, we don't necessarily need to write-lock `index.html`; we just need to make sure
that the processes run sequentially. So, we can instead use a mutex (via portalocker's
BoundedSemaphore class) which will generate the semaphores in `tempfile.gettempdir()` by
default. As a fallback, if a tempdir doesn't exist, the PWD will be used. (See
https://docs.python.org/3/library/tempfile.html#tempfile.gettempdir)

This will avoid the issue of trying to lock a file on an NFS-mounted drive.