Update check_image_kind-based colormapping
In my hasty WIP, I found that some test softsegs only had a ~93% percentage of binary voxels, due to a really tight crop around the spinal cord. To fix this, I reduced the binary percentage to 90%.

However, I'm not sure that this is truly necessary. For the deepseg SC, since it's... well... deepseg, I think we can safely assume that we'll only ever want to display either binary segs or softsegs. In that case, I think we can just assume that if an image is not a binary segmentation, then we should automatically use the softseg colormap.

So, I've reverted the threshold change and added this new condition.