`qc2.py`: Fix aspect ratio bug
The previous qc.py code adapted the aspect ratio to the zooms of the image. However, this causes
a bug due to the prior slicewise resampling to 0.6mm isotropic -- sometimes the resampled
resolution won't be exactly 0.6 (e.g. 0.594...), meaning the ratio will be a little off, and the
final image will be stretched/skewed a bit.

Since we know in advance that we're resampling to an isotropic resolution, we can just hardcode
1.0.