`sct_deepseg`: Restrict `-c` behavior to 'seg_tumor-edema'
The model 'seg_tumor-edema-cavity_t1-t2' is different compared to the rest of the `sct_deepseg`
models, in that it requires 2 input images, each with different contrasts. In order to
accommodate for this model, `sct_deepseg` contains some logic specific to the `-c` option
that is only relevant for this model (input ordering based on contrast).

However, this logic is universally applied to all models, even when `-c` is not relevant. This
restricts our ability to use `-c` for other means (for example, multiplying PSIR images by -1).

So, to enable `-c` to be used for other means, we limit the behavior specifically to
'seg_tumor-edema-cavity_t1-t2'.