`sct_deepseg`: Add inversion step for PSIR images
The logic goes like this:

-   no `-c`: Error is thrown.
- `-c psir`: Image is inverted.
- `-c stir`: Inversion is turned off.

I chose to make `-c` mandatory because I was concerned that users would miss the behavior if it
were implicit, for example:
- Invert by default, "opt-out" using `-c stir`: Users may accidentally invert STIR images.
- Do nothing by default, "opt-in" using `-c psir`: Users may forget to invert PSIR images.

----

There is a bit of awkwardness re: *where* to do the inversion. The issue is that the `-c` flag
is checked in `sct_deepseg.py` (due to argparsing), but the tmpdir is created two levels deeper
in the function `segment_nnunet`. So, I figured that I had two options:

1. Invert the image within `sct_deepseg.py` (by creating a new tmpdir and replacing the input
   filename)
2. Carry the result of `-c` into the function calls, then invert the image in the tmpdir.

I went with 1. because it keeps the inversion code isolated in a single location,
and it avoids a bit of extra clutter in the function signatures.