Isolate `index.html`-writing code into reusable script + replace `Lock` with Semaphore-based `mutex` (#4439)
This PR contains 2 key changes, both with the goal of better isolating
the part of the QC code that writes the report files to the QC
directory:

1. Separate the "import-JSON-into-HTML" steps into a standalone script
that can be run both as part of the QC-generation, but also by the user
to remove unwanted entries.
   - Fixes #4438.
- Another benefit of isolating this code is that it gives me a place to
document the limitations of client-side JavaScript, which I spent quite
a bit of time researching. :sweat_smile:
2. Use a mutex instead of locking `index.html`.
- The mutex uses `portalocker.BoundedSemaphore`, which still uses
`portalocker.Lock` under the hood, but with some additional helper code
(retries, timeouts, semaphore creation in a tempdir on the local
filesystem), which helps us to avoid reinventing the wheel.
- This allows us to limit the number of parallel processes independently
of the location of `index.html`, which avoids issues with writing the QC
reports to NFS-mounted drives. (Fixes #4423.)
- (I'm presuming that SCT itself won't be installed on an NFS-mounted
drive, only the data, which was the case for the above issue.)