Add (internal, semi-private) support for `sct_deepseg` GPU inference on Linux (#4421)
This PR adds support for GPU inference, controlled by:

1. A new option to `install_sct` to install GPU torch.
2. An environment variable called `SCT_USE_GPU` which, if set, will
trigger SCT to use `'cuda'` as the device of choice.

I've tried to set the `device` for all of the relevant inference
packages (`ivadomed`, `monai`, `nnunet`) but I may have missed something
-- it's a bit tricky because the methods for setting `cuda` vary from
package to package.

The steps to use this PR are as follows:

> - Clone SCT to your home folder.
> - Install SCT using the `-g` flag (e.g. `./install_sct -iyg`)
> - Make sure that `sct_check_dependencies` displays the available GPUs.
> - Set `CUDA_VISIBLE_DEVICES` to one of the GPU IDs (e.g. I booked GPU
2 on rosenberg, so I set `CUDA_VISIBLE_DEVICES=2`). Then, re-run
`sct_check_dependencies` and make sure that the the "number of GPUs
available to PyTorch" drops to 1.
> - Note: Since PyTorch can only see the GPUs that are visible, they
will start with ID 0 internally to PyTorch, even though the "real" ID
for this GPU was 2. There's [not much we can do about
that?](https://discuss.pytorch.org/t/how-to-change-the-default-device-of-gpu-device-ids-0/1041/10)
> - Set `SCT_USE_GPU=1` to turn on GPU inference.
> - For now, this acts as a "hidden" switch to turn on GPU usage, only
for internal lab folks to know about. ;)
> - Run a sample command (e.g. `SCT_USE_GPU=1 sct_deepseg -task
seg_sc_lesion_t2w_sci -i t2.nii.gz`)

- This PR only adds support to `sct_deepseg`, and only for Linux. 
    - We may want to extend support to Windows as well.
- We may also want to consider installing `onnxruntime-gpu` as well,
however this currently conflicts with the `ivadomed` package, which
explicitly specifies the CPU version of `onnxruntime`.
- As it stands, this is a "hidden" option. We may want to make this a
visible option in `sct_deepseg` if we feel ready to accept the
maintenance burden that comes with this.