Add QC report for `sct_deepseg` (#4446)
## Description

I created a general QC for
`sct_deepseg`, rather than a QC for just the contrast agnostic softseg.
There are a lot of different models to consider, but I think I've done
things in a way that supports all of them. :crossed_fingers:

I wanted to focus on the axial mosaic to begin with, because that's
what's likely most pressing for ISMRM/showing off the contrast agnostic
seg. But, ideally we'd port the sagittal mosaic over too for lesion
segmentations.

## Linked issues

Fixes #4436.
Fixes #3325.

---------

Co-authored-by: Mathieu Guay-Paquet <mathieu.guay-paquet@polymtl.ca>