`qc2.py`: Switch from two-tone colormap to transparency
Using a dark color for values <1.0 was making it much more difficult to QC.

This solution scales the transparency to the image values. If the segmentation == 1.0, then the
image will be opaque red (as it already was). But, as the values get closer to 0, the voxels
will get more and more transparent. This will prevent low values (e.g. 0.01) from showing up in
the QC report.