Switch from two-tone colormap to transparency for `sct_deepseg` QC (#4460)
## Description

This PR is a small amendment to #4446. Using a dark color for values
<1.0 was making it much more difficult to QC the soft perimeter of
softsegs.

This solution scales the transparency to the image values. If the
segmentation == 1.0, then the image will be opaque red (as it already
was). But, as the values get closer to 0, the voxels will get more and
more transparent. This will prevent low values (e.g. 0.01) from showing
up in the QC report.

## Linked issues

Fixes #4457.