Fixed formatting
Previously, when entering wrong label, STDOUT was:
At least one of the selected labels ([600]) is not available                 according to the label list from the text file in the atlas folder.

Now it looks like this:
At least one of the selected labels ([600]) is not available according to the label list from the text file in the atlas folder.