Run all tests in temporary directories (to preserve a clean copy of `sct_testing_data`) (#4474)
## Description

**Context:** SCT used to manage its own testing framework. This
framework was ported to `pytest` in
https://github.com/spinalcordtoolbox/spinalcordtoolbox/pull/3373.
However, many of the old tests used relative paths to input data, which
would fail unless the new pytest tests were run in `sct_testing_data`
(or the tests were rewritten).

At the time, in order to keep the diffs for
https://github.com/spinalcordtoolbox/spinalcordtoolbox/pull/3373 a
little more manageable, I opted to keep the relative paths but add a
fixture to run the tests in `sct_testing_data`. This was never meant to
be a long-term solution, as it meant that outputs to the PWD would
clutter `sct_testing_data`.

This PR is that final follow-up to properly rewrite all relative paths,
so that we can run the tests themselves in tempdirs, rather than
`sct_testing_data`. (This has the added benefit of rendering the data
integrity check actually useful now, since it shouldn't throw spurious
errors anymore.)

## Linked issues

Fixes #2911.