Fix tiny sagittal mosaic QC Report for large images with many slices (#4479)
## Description

Sagittal mosaics are a (relatively) new type of QC report, added in
https://github.com/spinalcordtoolbox/spinalcordtoolbox/pull/4102. They
are based off of mosaics of Axial slices. However, there is a key
difference: Axial slice mosaics are fixed at 30x30vox per axial slice,
whereas sagittal slice mosaics vary depending on the extents of the
sagittal segmentation.

In #4458, @NathanMolinier discovered an edge case: If the resolution of
the image is sufficiently large, and if the segmentation spans many
sagittal slices, then the mosaic can grow way too big in size. At that
point, it hits our height limitations for QC reports (both in matplotlib
canvas size, and CSS max-height), and thus the whole mosaic image gets
scaled down to fit.

So, this PR adjusts the height limitations to allow sagittal mosaics to
display in full size. However, this causes a new issue: Now the QC report is so 
large that it becomes unwieldy to actually view, thanks to the resampling step. 
(1.0mm -> 0.6mm causes the voxel size of the image to balloon.) So, to
counteract that, this PR adds a `-resample` option to allow the user to
vary the resampling value to change the zoom of the QC report.

## Linked issues

Fixes #4458.

---------

Co-authored-by: Mathieu Guay-Paquet <mathieu.guay-paquet@polymtl.ca>