`test_sct_extract_metric`: Add failing test for OOB -vert values
For data where an image is split into multiple chunks, you may want to call the same command
on multiple chunks spanning a range of vertlevels (e.g. -vert 1:12). But, not all chunks will
span all levels.

In this case, there is a bug where a table entry will be created for the slicegroup "()", i.e.
a group containing no slices. This is a useless row containing no data.

This commit adds a test that reproduces the condition and fails. As soon as the empty slicegroup
row is removed, the test will pass.