`aggregate_slicewise.py`: Skip `()` when making `agg_metric`
The `if slicegroups`/`if not slicegroups` check is duplicated, which I don't love. An alternative
could be to remove empty slicegroups from `slicegroups` entirely. But, then we wouldn't be able
to use `vertgroups[slicegroups.index(slicegroup)]`, since vertgroups and slicegroups would be
different sizes.

In general, this code really needs a refactor. See this already open issue:

- https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/4005