`sct_maths`: Replace `argv` parsing with `arguments.ordered_args`
Rather than accessing the parsed argument values from `arguments` (which will be overwritten when
the same argument is passed multiple times), we fetch the argument values from the new
`arguments.ordered_args`, which preserves the arguments (and their parsed values) in the order they
were passed.

One thing I don't love about this solution is that we still need to access the "global variables"
`arguments.shape` and `arguments.dim`, which was predicted by @mguaypaq in https://github.com/spinalcordtoolbox/spinalcordtoolbox/pull/4485#pullrequestreview-2069546042. I'll try to
puzzle out a better way of doing this...