`sct_maths`: Group `shape` and `dim` arguments with dilate/erode
This avoids us having to pass `arguments` to `apply_array_operation`, but I feel like I'd much
rather remove `-shape` and `-dim` entirely, and instead have these values passed directly to
`-dilate` and `-erode`. This would allow us to specify multiple different values for shape/dim, one
per call to `-dilate`/`-erode`. As it stands, these values are "global", which is a bit of a hindrance
for the user.