`image.py :: check_image_kind`: Rework `binary_most_common`
The previous logic assumed that seg/softseg would have 0.0 and 1.0 be the two most common
values. However, some softsegs will meet the "binary_percentage > 0.95" condition,
yet have 1.0 be its third or fourth most-common value (due to the prevalence of other values
in the range [0, 1]).

Because of this, I think we can rely just on `binary_percentage` and not `binary_most_common`.