Properly detect softsegs for QC when binary values are *not* the most common values. (#4494)
## Description

The code expects soft segmentations to be similar to binary
segmentations (0.0/1.0), but with some nonbinary values in the range
[0.0, 1.0] to represent the "soft" regions.

However, In #4491, the lesion softseg is low resolution, making the soft
perimeter very thick.

This means that, while 0.0 is the most common value, 1.0 is _not_ the
second-most common value, and the softseg heuristic fails. Upon
reflection, I think this "`binary_most_common`" heuristic is too strict;
it should be sufficient to just check the overall percentage of 0.0 and
1.0 voxels.

Additionally, this PR adds a condition to ensure that softsegs are in
the range `[0, 1]`, which should have been there from the start.

## Linked issues

Fixes #4491.