`shell.py`: Simplify overly-rigorous argparse `__init__`
The original changes were made when I was a fresh dev. I believe my goal was to try to
preserve the original `ArgumentParser` default arguments (necessitating *args, **kwargs)
while also overriding some of the default values of the parent class, leading me to:
https://stackoverflow.com/a/41623488

However, as Mathieu pointed out, we have very clearly documented usage of SCTArgumentParser
across all of our scripts, and none of them customize the ArgummentParser any further than
`description` and `epilog`. So, I think we can choose to only expose a limited subset of
the arguments, rather than preserving all of them, which removes the need for *args, *kwargs.

Since we are already modifying the `__init__` method in this PR for other reasons, it seems
fine to do this little refactor in the context of this PR. :)

Co-authored-by: Mathieu Guay-Paquet <mathieu.guay-paquet@polymtl.ca>