`test_utils_shell.py`: Fix `prog` behavior
The reason the test began to fail is that we eschewed the use of `inspect.stack`, and instead
hooked into the default Argparse behavior, which relies on `sys.argv`. So, now, the `prog`
is whatever script initially ran (pytest, PyCharm's JB test runner, etc.), rather than
the caller module.

This is easy enough to fix! And, it seems harmless enough to keep.