`monai.py`: Switch to "bilinear" resampling in pre/postprocessing
Note: I'm not sure if it's possible to have different interpolation modes for pre vs.
postprocessing. When I tried to use e.g. `spline 2` for preprocessing, but `bilinear` for
postprocessing, MONAI would throw an error related to a transform mismatch.

Is using `bilinear` for the preprocessing OK? Will it produce results that are as good as
`spline 2`? I would appreciate input from folks in the lab!