Add `matplotlib-inline` backend for compatibility with Jupyter notebooks (#4525)
## Description

`matplotlib-inline` is an optional package that comes with `jupyter` for
use with the '%matplotlib inline' directive. However, if a user runs
`jupyter` from a separate environment and sets '%matplotlib inline',
then SCT's environment will be missing this backend, and SCT scripts
will crash when trying to import matplotlib.

So, we install `matplotlib-inline` in our env by default to ensure that
SCT scripts can be run in Jupyter notebooks without error. (The package
is very light, and only adds one small additional sub-dependency
`traitlets`.)

## Linked issues

Fixes #4523.