Simplify calls to `np.einsum` (#4529)
This is a tiny refactor to simplify two lines of code in `types.py`,
since I recently had to dig through this file for debugging.

The reason is that:
- `np.transpose` is generally a little confusing to the reader, so it's
best to set up proper conventions up-front and avoid transposes if
possible. (It's not always possible, though.)
- `np.einsum` is a powerful tool, but it's a bit hard to read. Let's not
make it harder than it needs to be.
- `np.rollaxis` is massively confusing, so it's really best to avoid it.

And in the case of `np.einsum`, we have complete control over the
ordering of axes for the inputs and the outputs, so we really can avoid
transposes and rollaxes. (This is the equivalent of getting rid of
double negatives: it makes things easier to read!)

---------

Co-authored-by: Joshua Newton <joshuacwnewton@gmail.com>