Restrict `numpy<2` to avoid upstream errors (#4536)
## Description

Numpy just upgraded to 2.0.0 on Sunday, June 16th. Many of our
dependencies are not yet ready for numpy==2.0.0, so we are forced to
restrict `numpy<2` for now. (Numpy themselves also recommend this in
their error message.)

```
A module that was compiled using NumPy 1.x cannot be run in
NumPy 2.0.0 as it may crash. To support both 1.x and 2.x
versions of NumPy, modules must be compiled with NumPy 2.0.
Some module may need to rebuild instead e.g. with 'pybind11>=2.12'.

If you are a user of the module, the easiest solution will be to
downgrade to 'numpy<2' or try to upgrade the affected module.
We expect that some modules will need time to support NumPy 2.
```

## Linked issues

Partially addresses #4535. 

(We should probably keep the issue open as a reminder to periodically
check whether `numpy==2.0.0` is viable. I can take care of reporting
issues upstream to our more niche medical imaging packages to ensure
that this gets fixed in a timely manner.)

---------

Co-authored-by: Mathieu Guay-Paquet <mathieu.guay-paquet@polymtl.ca>