Make `sct_download_data` errors non-fatal during installations of SCT (#4484)
## Description

This PR follows the reasoning in
https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/4410#issuecomment-2112883978.
By allowing `sct_download_data` failures, we mitigate crashing the
entire SCT installation if a single dependency fails to install (e.g.
due to network/disk issues).

That said, it is still important for users to know that the installation
of dependencies has failed. So, we also add a new check to
`sct_check_dependencies` that checks if the default data dependencies
are installed. As a bonus, this means that the list of defaults no
longer has to be hardcoded inside of the individual install scripts.

## Linked issues

Fixes #4410.