Store operations in arguments.todo
The `StoreTodo` action can be used to store an operation in
`arguments.todo` as a triple of the form `(operation name,
parsed argument values, {})`. The `StoreTodoWithShapeDim` action
can be used the same way, and adds the current value of
`arguments.shape` and `arguments.dim` to the dictionary at the
end of the triple.