Avoid buggy version of `nnunetv2` that produces empty output predictions (#4564)
## Description

Given the comments in `requirements.txt`, it seems like 2.3.1 is the
only safe version of `nnunetv2` we can rely on for now.

Action item:

- [ ] Report this upstream and help debug why this issue is occurring in
the first place. (Maybe after SCT v6.4, since I have a backlog I need to
work through before I can spend time digging into a tricky upstream
issue.)

## Testing

Testing this is a bit tricky, because our test (on cropped input) still
works fine on 2.4.2? The problem only reveals itself on a full test
image.

Maybe we could write a test similar to the mouse test, and stick the
large image in the new testing dataset?

## Linked issues

Fixes #4556.