Adjust QC canvas size by physical resolution to avoid tiny QC with anisotropic T2* images (#4566)
## Description

Previously, we adjusted the canvas height by the image dimensions.
However, for highly anisotropic images with thick axial slices, we also
need to make sure to adjust by the pixel resolution. Otherwise, because
there are so few axial slices, the canvas will get squished to a very
small height.

## Linked issues

Fixes #4563.