`centerline/core.py`: Copy `im_seg` + reorient from RPI -> native
The reorientation to RPI was actually changing the image _outside_ of the centerline context, too.
Meaning, when we use `im_seg` later in the QC process, it stays in RPI, which produces a very
incorrect QC report.

To avoid mutating the input in the first place, we make a copy.

However, we still need to remember to reorient `im_seg` back to the native orientation, too,
because it gets used again later on in the `get_centerline` function. This is done for
the non-optic centerline methods, but was forgotten here.

Note that these changes imply that `get_centerline` was buggy for optic. (Is it possible that
this bug was affecting e.g. `sct_deepseg_sc`/`sct_propseg` and other users of `optic`?)