Set `edge` padding as default for contrast-agnostic monai model (#4576)
## Description

Currently, test-time preprocessing transforms for the monai models
(which, at the moment, is only the contrast-agnostic model), used
zero-padding during cropping and padding as in [these
lines](https://github.com/spinalcordtoolbox/spinalcordtoolbox/blob/master/spinalcordtoolbox/deepseg/monai.py#L176-L177).
However, @valosekj observed that zero-padding DWI images from the
`dcm-brno` dataset resulted in [shifted
predictions](https://github.com/valosekj/dcm-brno/issues/19#issuecomment-2237682249).
Interestingly, the shifts were only observed in the initial and final
slices, suggesting that this might be sub-optimal padding issue.

I experimented with different padding options as test-time
[here](https://github.com/sct-pipeline/contrast-agnostic-softseg-spinalcord/issues/113#issuecomment-2247689725)
and found that `edge`-padding fixed the issue with shifted predictions.

Hence, this PR updates the default padding (which is zero padding) to
`edge` padding. This change should not break anything as `edge` padding
is at least as good as the zero-padding (so it is not risky to make it
the default). Morever, I noticed that nnUNet has also started to
(subtly) make `edge` padding the default, supporting the changes in this
PR.

---------

Co-authored-by: Joshua Newton <joshuacwnewton@gmail.com>