`sct_analyze_lesion`: Move "total" summing to sub-method
Currently, we sum the totals at the same time we write the sheets to the file. But why? Doing it
this way requires some awkward checks for `#` in the sheet name.

If we were to instead sum the totals *in the sheet-specific method*, then we already know that we're
working in the correct context.

Additionally, this is necessary so that we can add additional columns *after* the total for the spinal
cord, white matter, gray matter, and other combined labels.