`sct_analyze_lesion`: Add `CombinedLabels` to ROI distribution
Luckily for us, this functionality was already implemented! We get to simplify the code here;
instead of hardcoding the ranges for GM/WM/etc., we can re-use the existing label ranges as defined
by info_label.txt. This also allows the user to specify custom "CombinedLabels", too.

Oddly, the old code used contiguous ranges. But, since we have the full lists of combined labels, we can
just use `in` rather than `min < val < max`, in case the CombinedLabel does not span a contiguous range.