`sct_analyze_lesion`: Add `-perslice` option
This was a really tricky refactor, because aggregating by vertebral levels is heavily
baked into `sct_analyze_lesion`'s functionality. There are mentions of 'vert' peppered
all-throughout the script. It took numerous surgical changes of (vert -> row) to avoid
a complete rewrite.

That said, big-picture "aggregating by vertebral levels" basically boils down to this:

- Iterate over vertebral levels.
- For each vertebral level, set all other voxels equal to 0, then compute percentages.

The general structure of this code still works if you substitute "vertebral level"
for slice; You can instead iterate over slices and still set all other voxels equal
to 0.

To do this, I use "rows" to represent either vertebral levels *or* slices, and keep
a separate variable "row_name" to track whether we have vertlevels or slices. This
mostly gets the job done.

Note: While the previous performance improvements help, this is still a slow process
for any image with a large number of slices:

- `-perslice 0`, no optimization: ~300s (~5m)
- `-perslice 0`, with optimization: ~50s (~1m)
- `-perslice 1`, no optimization: Did not finish (est. 3600s, or 1hr)
- `-perslice 1`, with optimization: ~600s (10m)

So, perhaps there is more optimization work to be done.