Add `-perslice` option (plus performance improvements to make it feasible) (#4588)
This was a tricky refactor, because `-perslice` goes against the default
behavior of `sct_analyze_lesion` (aggregating by vertebral levels), and
there are a lot of hardcoded `vert` variable names. It took quite a few
changes of (vert -> row) to avoid a complete rewrite.

That said, big-picture "aggregating by vertebral levels" basically boils
down to this:

- Iterate over vertebral levels.
- For each vertebral level, set all other voxels equal to 0, then
compute percentages.

The general structure of this code still works if you substitute
"vertebral level" for slice; You can instead iterate over slices and
still set all other voxels equal to 0. To do this, I used "rows" to
represent either vertebral levels *or* slices, and kept a separate
variable "row_name" to track whether we have vertlevels or slices. This
mostly gets the job done.

Note: While the previous performance improvements help, this is still a
slow process for any image with a large number of slices:

- `-perslice 0`, no optimization: ~300s (~5m)
- `-perslice 0`, with optimization: ~50s (~1m)
- `-perslice 1`, no optimization: Did not finish (est. 3600s, or 1hr)
- `-perslice 1`, with optimization: ~600s (10m)

So, perhaps there is more optimization work to be done.

It's a lot of calls to `np.sum`, `np.count_nonzero`, etc. I think this
is because we've got naive `for` loops over both slices (~500) and
tracts (~36). I'm sure there's some sort of smart numpy array operations
we could be doing, but it's hard to implement that without a full
rewrite.

It may also be worth just... skipping the slices that don't have lesion
values? With large images, this results in a large output spreadsheet
with many empty rows. But, if we exclude the empty rows, then perhaps we
can save on processing time as well.

Fixes #4567 
Fixes #2677

---------

Co-authored-by: Mathieu Guay-Paquet <mathieu.guay-paquet@polymtl.ca>