Silence `pandas` warning by explicitly using `copy` to remove view/copy ambiguity (#4596)
As per
https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/4591#issuecomment-2271973627:

> But we are not explicit about whether we want a "view" (i.e. changes
the original dataframe) or a "copy" (independent of the original
dataframe). Because we are not explicit, there is potential ambiguity,
hence why pandas throws this warning. To fix this warning, we just need
to be explicit about what we want, and there are multiple ways to do
this (not just `.loc`).

I think in this case, because we only ever access the newly-added
columns from df_temp and not from the original `tissue_bridges_df`
DataFrame, we don't need a "view", and so I think it should be fine to
denote df_temp as a copy:

Fixes #4591.