Support `CombinedLabels` from `info_label.txt` when computing lesion distributions with `-f` (#4584)
This PR amends two methods responsible for computing the distribution of
a lesion within a warped atlas (i.e. the `-f` option of
`sct_analyze_lesion`):

- `_measure_eachLesion_distribution`: Responsible for the
`lesion#Ndistribution` pages in the sheet. The values are "percentage of
the lesion that exists in a given ROI". Since the CombinedLabels are
just combined ROIs, we can simply sum the percentages.
- `_measure_totLesion_distribution`: Responsible for the
`ROI_occupied_by_lesion` page in the sheet. The values are "percentage
of the ROI taken up by _any_ lesion". This is a bit trickier, since the
ROI area is in the denominator (rather than the numerator). Thankfully,
it turns out that this functionality was already implemented!
- Before: Hardcoded "combined labels" that are identical to the entries
in `info_label.txt`.
    - After: Read the "combined labels" directly from `info_label.txt`.

The end result is that the user can now modify `info_label.txt` to add
their own custom labels, and they will get results for the pages
generated by `-f`.

I also took the time to solve a small issue re: column ordering. All it
took was a single call to `sorted`.

I wanted to solve #2677 while I was at it, since this was also requested
by the user. However, I ran into an issue that I previously observed in
https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/4394#issuecomment-1986482032:

> > [!NOTE]
> > As a side observation, I noticed that adding the `-f` option makes
`sct_analyze_lesion` _very_ slow on my personal laptop, even for a
single subject. There is no feedback in the output while the
template-related processing is running, too. This isn't a great user
experience, and I would love to look further into improving this after I
debug the numerical issues.

By switching from per-level aggregation (usually ~10-13 vert levels) to
per-slice aggregation (as many as 500 slices), the optimization issues
get _much_ worse. (As in, couldn't even finish processing worse.) So, to
properly implement this feature, it would require some sort of
optimization fix to improve the run time. At that point, I figured it
might be better to get this initial PR going, and follow-up with a
second fix afterwards.

Fixes https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/4567
Fixes https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/4568
Fixes #4587 (When #4588 is merged)
Fixes #2677 (When #4588 is merged)

---------

Co-authored-by: Mathieu Guay-Paquet <mathieu.guay-paquet@polymtl.ca>